csh installed: the hang was a 16-bit p_pid, not a signal bug
THE RECORDED DIAGNOSIS WAS TRUE AND THE INFERENCE FROM IT WAS WRONG.  d08fa1f
shipped csh built-but-not-installed with a stack sample naming pjwait asleep in
sigpause, and a session went into the signal layer on the strength of it.  The
signal machinery was correct throughout: a forty-line probe reproducing pjwait's
loop outside csh -- sighold, read the flag, sigpause -- worked on the first run,
which is what moved the search back into csh.

sh.proc.h declares `short p_pid'.  palloc stores what fork returned; pchild
compares it against what wait3 returns (sh.proc.c:51).  V8 pids wrapped at 30000
and a VAX short held every one; macOS goes to 99998.  Measured, both ends
printed side by side:

  palloc: recorded pid 45267
  pjwait: pp->p_pid    -20269      <- 0xB0D3 read as a signed short
  pchild: wait3 ->      45267
  pchild: NO MATCH for  45267

so PRUNNING is never cleared and pjwait waits forever for a child it has already
reaped.  Sixth member of the 16-bit-range table and the FIRST in a program's own
struct rather than behind a shim seam -- the identity half, like d_ino.  Safe to
widen by the one-end rule.  p_jobid goes with it (it is a pgrp, killpg takes it),
and sh.h's shpgrp/tpgrp/opgrp carry a SECOND defect: each has its address taken
for ioctl(TIOCGPGRP), which is `int *' in V8's header and in our shim, so the get
wrote four bytes into a two-byte global.

AND NO BEHAVIOURAL TEST CAN GUARD IT.  A freshly booted host hands out low pids,
so every "run a command" case passes against the broken shell on any CI runner.
Mutation measured worse than that: restoring the short fired THREE OF SIX value
cases, because what matters is the pid the CHILD gets, not the ceiling.  So the
load-bearing case is the field WIDTH, over a file set derived from the built
objects (which is what makes it correct about nfunc.c, that upstream's own OBJS
never names), plus a line printing how high this host's pids go.  The aimed
mutation is shpgrp -- job-control-only, no behavioural case touches it -- and it
fires the width case and nothing else.

The sweep it forced found a second instance: w.c's `short w_pid' copies
mproc.p_pid, which THIS PORT widened to int.  Unexercised (w exits at :469 with
No mem) and recorded as unexercised.

MOVING THE TREE TO ANOTHER MACHINE FOUND TWO MORE, in a tree that was green
minutes earlier:

  - tests/jail linked /bin/cat into $TMPDIR, and a hard link cannot cross a
    filesystem -- EXDEV once the tree sat on a second volume.  It also only ever
    tested the FIRST name.  Now links inside the rootfs (same filesystem by
    construction), tests both names, and asserts by INODE rather than size.
  - df HAD BEEN WRONG SINCE IT WAS WRITTEN.  dfree() takes the row's dev and dir
    from the mtab entry it matched by device number and took its NUMBERS from the
    caller's path.  In the jail stat("/") is $V8ROOT, so the `/' row is correctly
    labelled with the rootfs's volume and reported the host root's block count:
    1948455240 against 976557016.  The host-property trap in a PROGRAM rather
    than a test -- and tests/kmemu carried the same assumption from the other
    end, comparing one row against the host's figure for that row's device, which
    agrees exactly when the bug is invisible.  It checks every row by mount point
    now; replaying the defect names TWO bad rows where the old form saw one.

make test: 2473 passed, 0 failed across 17 suites.  Crash-probe floor unchanged
at 106 over 7685 invocations; csh adds none.  Installed commands 162 -> 163.

+++ v8/src/cmd/csh/sh.func.c
@@ -1005,7 +1005,7 @@ setlim(lp, limit)
 dosuspend()
 {
 	int old, ldisc;
-	short ctpgrp;
+	int ctpgrp;		/* PORT: sh.h's tpgrp reason */
 	int o;
 
 	if (loginsh)
